+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format of this change log follows the advice given at [Keep a CHANGELOG](htt
 - Geopattern generated program images are served via standard pluginfile.php
 - All images were moved to system context, program context will be used only for management access control and tenant separation
 - All tags were moved to system context
+- Moved favourites storage to system context
 
 ## [v4.5.10.01](https://github.com/mutms/moodle-tool_muprog/compare/mu-4.5.9-01...v4.5.10.01) - 2026-02-12